パスワードリセットページのコンポーネント分割完了
