+++ src/components/organisms/index.tsx
@@ -2,5 +2,6 @@ export { Banner } from './Banner';
 export { Header } from './Header';
 export { MoviesList } from './MoviesList';
 export { MyListItems } from './MyListItems';
+export { ResetForm } from './ResetForm';
 export { Trailer } from './Trailer';
 export { TwoColumnInfo } from './TwoColumnInfo';
