+++ src/components/pages/Reset.tsx
@@ -5,5 +5,15 @@ import { FC } from 'react';
 import { ResetScreen } from '../templates';
 
 export const Reset: FC = () => {
-  return <ResetScreen />;
+  return (
+    <ResetScreen
+      logo={'JACKETFLIX'}
+      pageTitle={'パスワードリセット'}
+      leadText={[
+        'メールアドレスを入力し、「リセットする」ボタンを押してください。',
+        <br key="leadTextBr" />,
+        '入力されたアドレス宛にパスワードリセット用のメールをお送りします。',
+      ]}
+    />
+  );
 };
